refactor: rename sync::sync to sync::sink
Clippy doesn't like nested module names.
